requirements.txt
python-dotenv==1.0.0
openai==1.12.0
tiktoken==0.3.3
pillow==10.2.0
requests
streamlit==1.24.0
msal==1.22.0
pytz==2021.3